Designed and finished the contact and about page.
Created a contact page that includes a mailto going to the classter submission. Created an about page that gives information about the MCAST mission statement and the team that manages MCAST. Started an FAQ page, while also updating previous pages and css style to fit with the newly added pages.