Finalized Web pages and style
Finished the FAQ page and adjusted previous pages to match with the newly created one. Style was also updated to suit all web pages. Assets/Images were moved to another folder to make submission more manageable easier to navigate. Once task 4 is done, all tasks and assets will be uploaded to the main branch in a folder called ''main'' with sub-folders in it for tasks 1,2 and 4 and a folder for assets.